+++ NEWS.md
@@ -1,3 +1,5 @@
+# conflicted (development version)
+
 # conflicted 1.1.0
 
 * New `conflicted_prefer_all()` and `conflicted_prefer_matching()` to