+++ NEWS.md
@@ -1,5 +1,7 @@
 # conflicted (development version)
 
+* `conflict_prefer_all()` and `conflict_prefer_matching()` are now much faster.
+
 * New `conflicts_prefer()` to easily declare multiple preferences at once:
   `conflicts_prefer(dplyr::filter, lubridate::week, ...)` (#82).