+++ NEWS.md
@@ -3,6 +3,10 @@
 * conflicted now understands `include.only` and `exclude` arguments that 
   you might have specified in `library()` (#84).
 
+* `conflict_prefer_all()` and `conflict_prefer_matching()` are now much faster.
+  And when `losers` is supplied, they only register the minimal necessary
+  number of conflicts.
+
 * New `conflicts_prefer()` to easily declare multiple preferences at once:
   `conflicts_prefer(dplyr::filter, lubridate::week, ...)` (#82).